Add IPaginatedResponse interface and PaginatedResponse implementation for pagination handling
