+++ Enhetsregisteret/AT.Common.Enhetsregisteret.Publish/AT.Common.Enhetsregisteret.Publish.csproj
@@ -10,7 +10,7 @@
     <Company>Arbeidstilsynet</Company>
     <Description>A http client which interfaces with enhetsregisteret (Brønnøysundregistrene)</Description>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <Version>3.0.0-alpha4</Version>
+    <Version>3.0.0-alpha5</Version>
     <PackageReadmeFile>README.md</PackageReadmeFile>
     <PackageReleaseNotes>(Package release notes are in CHANGELOG.md)</PackageReleaseNotes>
     <PackageProjectUrl>https://github.com/Arbeidstilsynet/dotnet-common</PackageProjectUrl>
@@ -41,6 +41,7 @@
       Include="..\..\shared\src\MemoryCachingHandler.cs"
       Link="Shared\MemoryCachingHandler.cs"
     />
+    <Compile Include="..\..\shared\src\Pagination.cs" Link="Shared\Pagination.cs" />
   </ItemGroup>
   <Target
     Name="InjectPackageReleaseNotesFromFile"
